+++ apps/web/src/utils/token-utils.ts
@@ -1,4 +1,4 @@
-import type { KISAccessTokenResponse } from '@/entities/stock';
+import type { KISAccessTokenResponse } from '@/features/stock/server';
 
 export class TokenUtils {
     /**
